backend/requirements.txt
flask
flask-cors
pymongo
werkzeug
bcrypt
python-dotenv
torch
torchvision
transformers
pillow
numpy
pymongo